Improve Elasticsearch queries
Lately we've been running into an issue where the scroll timeout
is met and where queries in general are just taking too long. These
commits aim to improve the situation by:

1. Setting a longer scroll, which starts at 30s and scales up by
the (square root) of the number of days in the query. This shouldn't
have a huge performance impact on the rest of the operations of ES
because the scan() helper function is supposed to clear scroll ids
as it finishes with them.

2. Query one index at a time, matching the index pattern given,
until no more data is found, then stop querying. Indices are queried
in reverse (chronological) order since most of our reports are done
by looking back at the last n days of data. (The first index is also
always checked in case of a configuration error with the index alias.)
Once data is found, the next time an index returns no data, we stop
trying to query any more indices.

3. Use "_doc" sorting. Per Elasticsearch docs, this can speed up
queries (especially scrolling queries) where neither score nor sort
is important, and neither is important for these reports. We also
make sure to explicitly tell Elasticsearch that we don't want a score.

4. Put all of the above in the BaseFilter, then refactored all of the
other classes to only add what's needed to the query options coming
out of the BaseFilter's get_query method. This reduces complexity
going forward and improves understanding of which scroll and other
query options are being used across all reports.