INF-1856 Allow arbitrary sort columns, use num jobs to sort most reports
Before this commit, the sort columns was hardcoded to "All CPU Hours" and
required hacky workarounds to change for different reports. This commit
allows each filter class to set its preferred sort column.

The impetus for this change was the request to have most of the
reports use the number of jobs as the sort column, which has also
been updated.