INF-1954 Add job unit metrics for CHTC report
A job unit is defined as:
1 CPU
4 GB Memory
4 GB Disk

And the number of job units a job "takes up" is the number of units
needed to fit the largest dimension of the job's resource requests,
e.g. a 2 CPU, 12 GB Memory, 4 GB Disk job will take 3 units because
of the memory request (12 GB / 4 GB).